💡 feat: separate tab logics
