+++ src/components/AppTooltip.vue
@@ -9,7 +9,7 @@ const show = ref(false)
     <slot></slot>
     <div
       v-if="show"
-      class="absolute opacity-90 w-[200px] z-10 p-1 bg-gray-500 text-white rounded shadow-lg text-xs"
+      class="absolute opacity-90 w-[150px] z-10 p-1 bg-gray-500 text-white rounded shadow-lg text-xs"
       style="bottom: 150%; left: 50%; transform: translateX(-50%);"
     >
       <slot name="content"></slot>
