+++ src/style.css
@@ -4,6 +4,7 @@ html, body {
 
 * {
   box-sizing: border-box;
+  font-family: 'Inter', sans-serif;
   font-family: 'Roboto Condensed', sans-serif;
 }
 
@@ -18,4 +19,14 @@ html, body {
   --vc-accent-700: #eb756e;
   --vc-accent-800: #e86057;
   --vc-accent-900: #E6554C;
+}
+
+::placeholder { /* Chrome, Firefox, Opera, Safari 10.1+ */
+  opacity: .5;
+}
+:-ms-input-placeholder { /* Internet Explorer 10-11 */
+  opacity: .5;
+}
+::-ms-input-placeholder { /* Microsoft Edge */
+  opacity: .5;
 }
